requirements.txt
pandas
numpy
nltk
faiss-cpu
sentence-transformers
rank_bm25